wechat_admin/requirements.txt
django==5.1.5
djangorestframework==3.15.2
django-simpleui==2024.1.1
django-cors-headers==4.3.1
django-import-export==3.3.7
djangorestframework-simplejwt==5.4.0
pillow==10.2.0
requests==2.31.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
